next refactor
