+++ app/src/main/java/edu/agh/susgame/front/managers/GameManager.kt
@@ -6,9 +6,9 @@ import edu.agh.susgame.dto.rest.model.PlayerId
 import edu.agh.susgame.dto.rest.model.PlayerREST
 import edu.agh.susgame.front.service.interfaces.GameService
 import edu.agh.susgame.front.ui.components.common.util.Coordinates
-import edu.agh.susgame.front.ui.components.common.graph.Edge
-import edu.agh.susgame.front.ui.components.common.graph.EdgeId
-import edu.agh.susgame.front.ui.components.common.graph.Path
+import edu.agh.susgame.front.ui.components.common.graph.edge.Edge
+import edu.agh.susgame.front.ui.components.common.graph.edge.EdgeId
+import edu.agh.susgame.front.ui.components.common.graph.edge.Path
 import edu.agh.susgame.front.ui.components.common.graph.node.Host
 import edu.agh.susgame.front.ui.components.common.graph.node.Node
 import edu.agh.susgame.front.ui.components.common.graph.node.NodeId
