+++ app/src/main/java/edu/agh/susgame/front/ui/components/game/components/drawers/NodeDrawer.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.unit.dp
 import edu.agh.susgame.R
 import edu.agh.susgame.dto.rest.model.PlayerId
 import edu.agh.susgame.front.managers.GameManager
-import edu.agh.susgame.front.ui.components.common.graph.PathBuilder
+import edu.agh.susgame.front.ui.components.common.graph.edge.PathBuilder
 import edu.agh.susgame.front.ui.components.common.graph.node.Host
 import edu.agh.susgame.front.ui.components.common.graph.node.Node
 import edu.agh.susgame.front.ui.components.common.graph.node.NodeId
