+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/createlobby/CreateLobbyView.kt
@@ -14,7 +14,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
-import edu.agh.susgame.front.ui.components.common.navigation.MenuRoute
+import edu.agh.susgame.front.ui.components.menu.navigation.MenuRoute
 import edu.agh.susgame.front.service.interfaces.LobbyService
 import edu.agh.susgame.front.ui.components.common.util.Translation
 import edu.agh.susgame.front.ui.components.common.theme.Header
