+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/lobby/elements/FailedToLoadComp.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
 import edu.agh.susgame.dto.rest.model.LobbyId
 import edu.agh.susgame.front.ui.components.common.util.Translation
-import edu.agh.susgame.front.ui.components.common.navigation.MenuRoute
+import edu.agh.susgame.front.ui.components.menu.navigation.MenuRoute
 
 @Composable
 fun FailedToLoadComp(lobbyId: LobbyId, navController: NavController) {
