+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/lobby/elements/LobbyComp.kt
@@ -27,7 +27,7 @@ import androidx.navigation.NavController
 import edu.agh.susgame.dto.rest.model.Lobby
 import edu.agh.susgame.dto.rest.model.PlayerNickname
 import edu.agh.susgame.front.ui.components.common.util.Translation
-import edu.agh.susgame.front.ui.components.common.navigation.MenuRoute
+import edu.agh.susgame.front.ui.components.menu.navigation.MenuRoute
 import edu.agh.susgame.front.service.interfaces.GameService
 import edu.agh.susgame.front.service.interfaces.LobbyService
 import edu.agh.susgame.front.ui.components.common.theme.Header
