+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/mainmenu/MainMenuView.kt
@@ -11,7 +11,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
-import edu.agh.susgame.front.ui.components.common.navigation.MenuRoute
+import edu.agh.susgame.front.ui.components.menu.navigation.MenuRoute
 import edu.agh.susgame.front.ui.components.common.util.Translation
 import edu.agh.susgame.front.ui.components.common.theme.Header
 import edu.agh.susgame.front.ui.components.menu.components.mainmenu.elements.MainMenuButton
