+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/searchlobby/SearchLobbiesView.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import edu.agh.susgame.dto.rest.model.Lobby
 import edu.agh.susgame.dto.rest.model.LobbyId
-import edu.agh.susgame.front.ui.components.common.navigation.MenuRoute
+import edu.agh.susgame.front.ui.components.menu.navigation.MenuRoute
 import edu.agh.susgame.front.service.interfaces.LobbyService
 import edu.agh.susgame.front.ui.components.common.util.Translation
 import edu.agh.susgame.front.ui.components.common.theme.Header
