+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/searchlobby/elements/LobbyRow.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import edu.agh.susgame.dto.rest.model.Lobby
-import edu.agh.susgame.front.ui.components.common.navigation.MenuRoute
+import edu.agh.susgame.front.ui.components.menu.navigation.MenuRoute
 import edu.agh.susgame.front.ui.components.common.theme.PaddingM
 
 @Composable
